requirements.txt
pandas==1.5.3
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-storage==2.12.0
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0
pyarrow==14.0.2 